Move request functions to the proper API interface.
